Use `PopenSpawn` in `check50.run`.
`pexpect.spawn` is not available on Windows as pseudo terminals are not
available, which the `pexpect` library uses.

This commit adds another method to `check50.run`, `isalive` to check if
the child process is still running.

The `check50.run.kill` will now wait until the child process exits after
sending the SIGKILL.

Also, this class now handles the killing of child processes if output
is mismatched(only in stdin when expecting a prompt) or timeout is
raised.