Wait until checks leave their temporary directory.
Windows does not allow deletion of folders with processes still in it.

This commit adds `.exit()` when compiling YAML checks which use `run`
key and do not have an `exit` key.

Adding an `.exit()` (which is the `check50.run.exit()`) will make the
check's parent process wait until the check process exits (or wait until
timeout, then the child process will be killed).

The pre-compiled checks in the testing suite are also modified within
this commit.